Moving Delete/Uninstall tasks to 'reponses' package
